% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{codes.default}
\alias{codes.default}
\title{Extract Codes by Default from an Object}
\usage{
\method{codes}{default}(x, simplify = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{simplify}{whether to convert length one list to vector}

\item{...}{passed arguments}
}
\value{
list, or vector if simplify = TRUE
}
\description{
Extracts codes from an object using the default method.
}

