% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decode.data.frame}
\alias{decode.data.frame}
\title{Decode Data Frame.}
\usage{
\method{decode}{data.frame}(x, ...)
}
\arguments{
\item{x}{inherits data.frame}

\item{...}{ignored}
}
\value{
same class as x
}
\description{
Decodes a data.frame. Calls as_factor() for each column with an encoded guide attribute.
}
