% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decode.default}
\alias{decode.default}
\title{Decode an Object by Default}
\usage{
\method{decode}{default}(x, encoding = NULL, ...)
}
\arguments{
\item{x}{object}

\item{encoding}{length one character that is itself encoded}

\item{...}{passed arguments}
}
\value{
factor
}
\description{
Decodes an object using the default method. Typically \code{x} is a character vector containing codes that can be extracted from \code{encoding}. Corresponding decodes are returned as a factor with levels of unique decodes. If \code{encoding} is NULL, it is replaced with an encoding such that levels and labels are both \code{unique(x)}. Duplicate codes are ignored.  Duplicate decodes are collapsed (combined to a single level).
}
\seealso{
Other decode: \code{\link{as_factor}},
  \code{\link{decode.data.frame}}, \code{\link{decode}}
}
