% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decode}
\alias{decode}
\title{Decode an Object}
\usage{
decode(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Decodes an object.  Default method supplied.
}
\seealso{
\code{\link{decode.default}}

Other decode: \code{\link{as_factor}},
  \code{\link{decode.data.frame}},
  \code{\link{decode.default}}
}
