% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decodes}
\alias{decodes}
\title{Extract Deodes from an Object}
\usage{
decodes(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts decodes from an object. Default method is supplied.
}

