% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode.character}
\alias{encode.character}
\title{Encode Character.}
\usage{
\method{encode}{character}(x, labels = NULL, sep = NULL, ...)
}
\arguments{
\item{x}{object}

\item{labels}{same length as x if supplied}

\item{sep}{a single character not present in x or labels}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Encodes character.
}
