% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode.default}
\alias{encode.default}
\title{Encode Default.}
\usage{
\method{encode}{default}(x, labels = NULL, ...)
}
\arguments{
\item{x}{object}

\item{labels}{same length as x if supplied}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Encodes using default method:  coerces to character and and encodes the result.
}

