% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode.list}
\alias{encode.list}
\title{Encode a List}
\usage{
\method{encode}{list}(x, labels = NULL, ...)
}
\arguments{
\item{x}{object}

\item{labels}{same length as x if supplied}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Encodes a list.
}

