% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encoded}
\alias{encoded}
\title{Check If Object is Encoded}
\usage{
encoded(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Checks if object is encoded.
}
\seealso{
\code{\link{encoded.default}}
}
