% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{codes}
\alias{codes}
\title{Extract Codes from an Object}
\usage{
codes(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts Codes from an object. Default method is supplied.
}

