% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decode.default}
\alias{decode.default}
\title{Decode an Object by Default}
\usage{
\method{decode}{default}(x, encoding = "", ...)
}
\arguments{
\item{x}{object}

\item{encoding}{length one character that is itself encoded}

\item{...}{passed arguments}
}
\value{
factor
}
\description{
Decodes an object using the default method. Typically \code{x} is a character vector containing codes that can be extracted from \code{encoding}. Corresponding decodes are returned as a factor.
}
