% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode.character}
\alias{encode.character}
\title{Encode Character.}
\usage{
\method{encode}{character}(x, labels = NULL, sep = NULL, ...)
}
\arguments{
\item{x}{object}

\item{labels}{same length as x if supplied}

\item{sep}{a single character not present in x or labels}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Encodes character.  If \code{sep} is NULL, it is replaced with the first of these that is not otherwise present in the result: /|:\\~!@#$%^&*?. (including dot).
}
\seealso{
Other encode: \code{\link{encode.default}},
  \code{\link{encode.list}}, \code{\link{encode}}
}
