% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encoded.default}
\alias{encoded.default}
\title{Check If Default Object is Encoded}
\usage{
\method{encoded}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Checks if object is encoded, using default methodology. Always returns logical, telling whether the corresponding element represents an encoding of levels and labels. Objects with zero length give \code{FALSE}.
}
\seealso{
Other encoded: \code{\link{encoded}}
}
